refactor: split jobs module into multiple files
