+++ ui/jobs/job/sch.js
@@ -1,0 +1,84 @@
+import { kAbility } from '../constants.js';
+
+export function setupSch(bars) {
+  const aetherflowStackBox = bars.addResourceBox({
+    classList: ['sch-color-aetherflow'],
+  });
+
+  const fairyGaugeBox = bars.addResourceBox({
+    classList: ['sch-color-fairygauge'],
+  });
+
+  const bioBox = bars.addProcBox({
+    id: 'sch-procs-bio',
+    fgColor: 'sch-color-bio',
+  });
+
+  const aetherflowBox = bars.addProcBox({
+    id: 'sch-procs-aetherflow',
+    fgColor: 'sch-color-aetherflow',
+  });
+
+  const lucidBox = bars.addProcBox({
+    id: 'sch-procs-luciddreaming',
+    fgColor: 'sch-color-lucid',
+  });
+
+  bars.jobFuncs.push((jobDetail) => {
+    const aetherflow = jobDetail.aetherflowStacks;
+    const fairygauge = jobDetail.fairyGauge;
+    const milli = Math.ceil(jobDetail.fairyMilliseconds / 1000);
+    aetherflowStackBox.innerText = aetherflow;
+    fairyGaugeBox.innerText = fairygauge;
+    const f = fairyGaugeBox.parentNode;
+    if (jobDetail.fairyMilliseconds !== 0) {
+      f.classList.add('bright');
+      fairyGaugeBox.innerText = milli;
+    } else {
+      f.classList.remove('bright');
+      fairyGaugeBox.innerText = fairygauge;
+    }
+
+    // dynamically annouce user depends on their aetherflow stacks right now
+    aetherflowBox.threshold = bars.gcdSpell() * (aetherflow || 1) + 1;
+
+    const p = aetherflowStackBox.parentNode;
+    const s = parseFloat(aetherflowBox.duration || 0) - parseFloat(aetherflowBox.elapsed);
+    if (parseFloat(aetherflow) * 5 >= s) {
+      // turn red when stacks are too much before AF ready
+      p.classList.add('too-much-stacks');
+    } else {
+      p.classList.remove('too-much-stacks');
+    }
+  });
+
+  bars.abilityFuncMap[kAbility.Biolysis] = () => {
+    bioBox.duration = 0;
+    bioBox.duration = 30;
+  };
+  bars.abilityFuncMap[kAbility.Bio] = () => {
+    bioBox.duration = 0;
+    bioBox.duration = 30;
+  };
+  bars.abilityFuncMap[kAbility.Bio2] = () => {
+    bioBox.duration = 0;
+    bioBox.duration = 30;
+  };
+  bars.abilityFuncMap[kAbility.Aetherflow] = () => {
+    aetherflowBox.duration = 0;
+    aetherflowBox.duration = 60;
+    aetherflowStackBox.parentNode.classList.remove('too-much-stacks');
+  };
+  bars.abilityFuncMap[kAbility.LucidDreaming] = () => {
+    lucidBox.duration = 0;
+    lucidBox.duration = 60;
+  };
+
+  bars.statChangeFuncMap['SCH'] = () => {
+    bioBox.valuescale = bars.gcdSpell();
+    bioBox.threshold = bars.gcdSpell() + 1;
+    aetherflowBox.valuescale = bars.gcdSpell();
+    lucidBox.valuescale = bars.gcdSpell();
+    lucidBox.threshold = bars.gcdSpell() + 1;
+  };
+}
