+++ ui/jobs/job/whm.js
@@ -1,0 +1,101 @@
+import EffectId from '../../../resources/effect_id.js';
+import { kAbility } from '../constants.js';
+
+export function setupWhm(bars) {
+  const lilyBox = bars.addResourceBox({
+    classList: ['whm-color-lily'],
+  });
+  const lilysecondBox = bars.addResourceBox({
+    classList: ['whm-color-lilysecond'],
+  });
+
+  const diaBox = bars.addProcBox({
+    id: 'whm-procs-dia',
+    fgColor: 'whm-color-dia',
+  });
+  const assizeBox = bars.addProcBox({
+    id: 'whm-procs-assize',
+    fgColor: 'whm-color-assize',
+  });
+  const lucidBox = bars.addProcBox({
+    id: 'whm-procs-lucid',
+    fgColor: 'whm-color-lucid',
+  });
+
+  // BloodLily Gauge
+  const stacksContainer = document.createElement('div');
+  stacksContainer.id = 'whm-stacks';
+  bars.addJobBarContainer().appendChild(stacksContainer);
+  const bloodlilyContainer = document.createElement('div');
+  bloodlilyContainer.id = 'whm-stacks-bloodlily';
+  stacksContainer.appendChild(bloodlilyContainer);
+  const bloodlilyStacks = [];
+  for (let i = 0; i < 3; ++i) {
+    const d = document.createElement('div');
+    bloodlilyContainer.appendChild(d);
+    bloodlilyStacks.push(d);
+  }
+
+  bars.jobFuncs.push((jobDetail) => {
+    const lily = jobDetail.lilyStacks;
+    // bars milliseconds is countup, so use floor instead of ceil.
+    const lilysecond = Math.floor(jobDetail.lilyMilliseconds / 1000);
+
+    lilyBox.innerText = lily;
+    if (lily === 3)
+      lilysecondBox.innerText = '';
+    else
+      lilysecondBox.innerText = 30 - lilysecond;
+
+    const bloodlilys = jobDetail.bloodlilyStacks;
+    for (let i = 0; i < 3; ++i) {
+      if (bloodlilys > i)
+        bloodlilyStacks[i].classList.add('active');
+      else
+        bloodlilyStacks[i].classList.remove('active');
+    }
+
+    const l = lilysecondBox.parentNode;
+    if ((lily === 2 && 30 - lilysecond <= 5) || lily === 3)
+      l.classList.add('full');
+    else
+      l.classList.remove('full');
+  });
+
+  bars.abilityFuncMap[kAbility.Aero] = () => {
+    diaBox.duration = 0;
+    diaBox.duration = 18 + 1;
+  };
+  bars.abilityFuncMap[kAbility.Aero2] = () => {
+    diaBox.duration = 0;
+    diaBox.duration = 18 + 1;
+  };
+  bars.abilityFuncMap[kAbility.Dia] = () => {
+    diaBox.duration = 0;
+    diaBox.duration = 30;
+  };
+  bars.abilityFuncMap[kAbility.Assize] = () => {
+    assizeBox.duration = 0;
+    assizeBox.duration = 45;
+  };
+  bars.abilityFuncMap[kAbility.LucidDreaming] = () => {
+    lucidBox.duration = 0;
+    lucidBox.duration = 60;
+  };
+
+  bars.gainEffectFuncMap[EffectId.PresenceOfMind] = () => {
+    bars.presenceOfMind = 1;
+  };
+  bars.loseEffectFuncMap[EffectId.PresenceOfMind] = () => {
+    bars.presenceOfMind = 0;
+  };
+
+  bars.statChangeFuncMap['WHM'] = () => {
+    diaBox.valuescale = bars.gcdSpell();
+    diaBox.threshold = bars.gcdSpell() + 1;
+    assizeBox.valuescale = bars.gcdSpell();
+    assizeBox.threshold = bars.gcdSpell() + 1;
+    lucidBox.valuescale = bars.gcdSpell();
+    lucidBox.threshold = bars.gcdSpell() + 1;
+  };
+}
